module-01-intro-to-ai-llms/requirements.txt
# Minimal deps for Module 1 demos (tokenizers, embeddings, basic LLM calls).
# Uncomment sentence-transformers for offline embedding demos.

openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
# sentence-transformers>=2.7.0
